+++ Assets/Common/FrostyMods.Common.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "FrostyMods.Core",
+    "name": "FrostyMods.Common",
     "rootNamespace": "",
     "references": [],
     "includePlatforms": [],
